Refactor bottom nav view model using use cases
