+++ feature/main/src/main/java/com/quare/blitzsplit/main/presentation/component/navbar/BottomNavBarItem.kt
@@ -6,26 +6,25 @@ sealed class BottomNavBarItem(
     val title: String,
     val icon: Int,
     val screen: BottomNavScreen,
-    open val isSelected: Boolean,
 ) {
+
+    abstract val isSelected: Boolean
+
     data class Groups(override val isSelected: Boolean) : BottomNavBarItem(
         title = "Grupos",
         screen = BottomNavScreen.GROUPS,
         icon = R.drawable.groups_icon,
-        isSelected = isSelected,
     )
 
     data class Contacts(override val isSelected: Boolean) : BottomNavBarItem(
         title = "Contatos",
         screen = BottomNavScreen.CONTACTS,
         icon = R.drawable.contacts_icon,
-        isSelected = isSelected,
     )
 
     data class Activity(override val isSelected: Boolean) : BottomNavBarItem(
         title = "Atividade",
         screen = BottomNavScreen.ACTIVITY,
         icon = R.drawable.activity_icon,
-        isSelected = isSelected,
     )
 }
