+++ feature/main/src/main/java/com/quare/blitzsplit/main/presentation/component/navbar/BottomNavBarViewModel.kt
@@ -2,6 +2,8 @@ package com.quare.blitzsplit.main.presentation.component.navbar
 
 import androidx.lifecycle.ViewModel
 import com.quare.blitzplit.component.navbar.BottomNavigationItem
+import com.quare.blitzsplit.main.domain.usecase.GetBottomBarFirstStateUseCase
+import com.quare.blitzsplit.main.domain.usecase.GetItemsUpdatedUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -9,41 +11,26 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
 
 @HiltViewModel
-class BottomNavBarViewModel @Inject constructor() : ViewModel() {
+class BottomNavBarViewModel @Inject constructor(
+    getBottomBarFirstState: GetBottomBarFirstStateUseCase,
+    private val getItemsUpdated: GetItemsUpdatedUseCase
+) : ViewModel() {
 
     private val _state: MutableStateFlow<BottomNavState> = MutableStateFlow(
-        BottomNavState(
-            selectedScreen = BottomNavScreen.GROUPS,
-            items = listOf(
-                BottomNavBarItem.Groups(isSelected = true),
-                BottomNavBarItem.Contacts(isSelected = false),
-                BottomNavBarItem.Activity(isSelected = false)
-            ).map { item ->
-                BottomNavigationItem(
-                    title = item.title,
-                    icon = item.icon,
-                    isSelected = item.isSelected,
-                    onClick = { onSelectItem(item) }
-                )
-            }
-        )
-
+        getBottomBarFirstState(::onSelectItem)
     )
 
     val state: StateFlow<BottomNavState> = _state
 
     private fun onSelectItem(clickedItem: BottomNavBarItem) {
-        _state.update {
-            it.copy(
-                items = it.items.map { item ->
-                    item.getUpdated(clickedItem)
-                },
+        _state.update { bottomNavState ->
+            bottomNavState.copy(
+                items = getItemsUpdated(
+                    items = bottomNavState.items,
+                    clickedItem = clickedItem
+                ),
                 selectedScreen = clickedItem.screen
             )
         }
     }
-
-    private fun BottomNavigationItem.getUpdated(
-        clickedItem: BottomNavBarItem,
-    ): BottomNavigationItem = copy(isSelected = title == clickedItem.title)
 }
